refactor util to services
